requirements.txt
azure-ai-ml
azure-ai-projects
azure-identity
openai
PyJWT
semantic-kernel